refactor: extract profile data refresh from App
